requirements.txt
streamlit
tensorflow
numpy
opencv-python-headless
pillow
yolo
opencv-python 
pillow 
numpy
ultralytics
